moved comments folder to components
