+++ react-app/src/components/Comments/Comments.css
@@ -1,3 +1,100 @@
+.comment-two-columns {
+  gap: 20px;
+}
+
+.new-comment-wrapper {
+  background-color: rgb(242, 242, 242);
+  height: 40px;
+  align-items: center;
+  margin-bottom: 10px;
+}
+
+.new-comment-placeholder {
+  width: 40px;
+  height: 40px;
+  background-image: linear-gradient(135deg, #e6846e, #70929c);
+}
+
+.new-comment-form {
+  border: 1px solid rgb(229, 229, 229);
+  height: 40px;
+  align-items: center;
+  justify-content: flex-end;
+  padding: 5px;
+  width: 100%;
+}
+
+.comment-input {
+  width: 100%;
+  height: 100%;
+  font-size: 14px;
+  font-weight: 500;
+  padding: 5px;
+  line-height: 1.4;
+  border: 1px solid #e5e5e5;
+  color: #333;
+  cursor: pointer;
+}
+
+.comment-input:focus {
+  outline: none;
+}
+
+.user-badge .avatar {
+  height: 120px;
+  width: 120px;
+}
+
+.song-description {
+  min-height: 150px;
+  overflow: auto;
+  width: 100%;
+}
+
+.song-comments-list {
+  width: 100%;
+}
+
+.song-comments-list * {
+  font-family: Lucida Grande, Lucida Sans Unicode, Lucida Sans, Garuda, Verdana,
+    Tahoma, sans-serif;
+  font-size: 12px;
+  line-height: 1.4;
+}
+
+.comments-count {
+  padding-bottom: 7px;
+  border-bottom: 1px solid #f2f2f2;
+  /* width: 100%; */
+  align-items: flex-end;
+}
+
+.comments-count-text {
+  font-size: 14px;
+  line-height: 1.3;
+  color: #999;
+}
+
+.no-comments {
+  width: 100%;
+  justify-content: center;
+  align-items: center;
+  height: 500px;
+  gap: 10px;
+}
+
+.no-comments h3 {
+  font-size: 18px;
+  font-weight: 100;
+  color: var(--color-border-dark);
+}
+
+.no-comments h4 {
+  font-size: 16px;
+  font-weight: 100;
+  color: var(--color-info);
+}
+
 .comment-card {
   padding: 10px 10px 10px 0;
   /* justify-content: space-between; */
