--- react-app/src/components/PlaylistFolders/SinglePlaylist/Playlist.css
@@ -1,23 +1,0 @@
-.playlist-tile {
-  width: 150px;
-  margin-right: 50px;
-}
-
-.playlist-summary {
-  justify-content: center;
-  align-items: center;
-  width: 110px;
-  height: 110px;
-  background-color: rgba(0, 0, 0, 0.8);
-  color: #fff;
-  box-sizing: border-box;
-  border-radius: 50%;
-  text-align: center;
-  text-transform: uppercase;
-  font-size: 12px;
-  gap: 5px;
-}
-
-.playlist-summary h1 {
-  font-size: 2.5em;
-}
