+++ react-app/src/components/PlaylistFolders/SinglePlaylist/Songs.css
@@ -1,3 +1,22 @@
+.playlist-summary {
+  justify-content: center;
+  align-items: center;
+  width: 110px;
+  height: 110px;
+  background-color: rgba(0, 0, 0, 0.8);
+  color: #fff;
+  box-sizing: border-box;
+  border-radius: 50%;
+  text-align: center;
+  text-transform: uppercase;
+  font-size: 12px;
+  gap: 5px;
+}
+
+.playlist-summary h1 {
+  font-size: 2.5em;
+}
+
 .playlist-two-columns {
   gap: 20px;
 }
