--- react-app/src/pages/SingleSong.css
@@ -1,171 +1,0 @@
-.single-song-banner {
-  background: linear-gradient(135deg, rgb(185, 172, 182) 0%, rgb(107, 108, 108) 100%);
-  position: relative;
-  margin: 0 var(--x-margin);
-  align-items: center;
-  padding: 30px 20px 20px 30px;
-  gap: 30px;
-  height: 390px;
-}
-
-.single-song-banner-left {
-  width: 100%;
-  height: 100%;
-  justify-content: space-between;
-}
-
-.single-song-banner-top-left {
-  align-items: flex-start;
-  width: 100%;
-  gap: 20px;
-}
-
-.song-page-play {
-  font-size: 3em;
-  color: var(--color-primary);
-  transition: color 0.15s;
-}
-
-.song-page-play:hover {
-  color: rgb(255, 47, 0);
-}
-
-.single-song-banner-info {
-  width: 100%;
-  justify-content: space-between;
-}
-
-.single-song-banner-info time {
-  color: white;
-  font-weight: 200;
-}
-
-.single-song-banner-ident {
-  gap: 5px;
-}
-
-.single-song-banner-ident h2,
-.single-song-banner-ident p {
-  background-color: var(--color-border-dark);
-  color: white;
-  padding: 4px 8px;
-  font-weight: 300;
-}
-
-.song-image {
-  width: 340px;
-  height: 340px;
-}
-
-.single-song-secondary {
-  gap: 30px;
-  padding-top: 20px;
-}
-
-.single-song-main {
-  width: 70%;
-}
-
-.song-button-group {
-  padding: 0 0 12px;
-  border-bottom: 1px solid grey;
-  width: 100%;
-  gap: 5px;
-}
-
-.comment-two-columns {
-  gap: 20px;
-}
-
-.new-comment-wrapper {
-  background-color: rgb(242, 242, 242);
-  height: 40px;
-  align-items: center;
-  margin-bottom: 10px;
-}
-
-.new-comment-placeholder {
-  width: 40px;
-  height: 40px;
-  background-image: linear-gradient(135deg, #e6846e, #70929c);
-}
-
-.new-comment-form {
-  border: 1px solid rgb(229, 229, 229);
-  height: 40px;
-  align-items: center;
-  justify-content: flex-end;
-  padding: 5px;
-  width: 100%;
-}
-
-.comment-input {
-  width: 100%;
-  height: 100%;
-  font-size: 14px;
-  font-weight: 500;
-  padding: 5px;
-  line-height: 1.4;
-  border: 1px solid #e5e5e5;
-  color: #333;
-  cursor: pointer;
-}
-
-.comment-input:focus {
-  outline: none;
-}
-
-.user-badge .avatar {
-  height: 120px;
-  width: 120px;
-}
-
-.song-description {
-  min-height: 150px;
-  overflow: auto;
-  width: 100%;
-}
-
-.song-comments-list {
-  width: 100%;
-}
-
-.song-comments-list * {
-  font-family: Lucida Grande, Lucida Sans Unicode, Lucida Sans, Garuda, Verdana,
-    Tahoma, sans-serif;
-  font-size: 12px;
-  line-height: 1.4;
-}
-
-.comments-count {
-  padding-bottom: 7px;
-  border-bottom: 1px solid #f2f2f2;
-  /* width: 100%; */
-  align-items: flex-end;
-}
-
-.comments-count-text {
-  font-size: 14px;
-  line-height: 1.3;
-  color: #999;
-}
-
-.no-comments {
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-  height: 500px;
-  gap: 10px;
-}
-
-.no-comments h3 {
-  font-size: 18px;
-  font-weight: 100;
-  color: var(--color-border-dark);
-}
-
-.no-comments h4 {
-  font-size: 16px;
-  font-weight: 100;
-  color: var(--color-info);
-}
