+++ react-app/src/pages/SingleSong.js
@@ -15,7 +15,6 @@ import CopyLinkButton from "../components/Buttons/CopyLinkButton";
 import EditButton from "../components/Buttons/EditButton";
 import DropdownButton from "../components/Buttons/DropdownButton";
 import EditSongForm from "../components/SongFolders/EditSongForm";
-import "./SingleSong.css";
 
 const ButtonGroup = ({ song }) => {
   const dispatch = useDispatch();
